feat: setup details pane tab switching
