+++ src/routes/problem/components/details_pane/components/description/description.scss
@@ -1,0 +1,121 @@
+.problem {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  justify-content: space-between;
+  padding: 16px;
+
+  &_details {
+    overflow-y: auto;
+  }
+
+  &_title {
+    font-size: 24px;
+    font-weight: 700;
+    margin-bottom: 8px;
+  }
+
+  &_metadata {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 8px;
+    margin-bottom: 24px;
+  }
+
+  &_description {
+    margin-bottom: 48px;
+
+    h1,
+    h2,
+    h3,
+    h4 {
+      font-weight: 600;
+      margin: 18px 0 12px;
+    }
+
+    p {
+      margin-bottom: 12px;
+    }
+  }
+
+  &_examples {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    margin-bottom: 48px;
+
+    .example_heading {
+      font-weight: 700;
+      margin-bottom: 6px;
+    }
+
+    .example_fields_container {
+      display: flex;
+      flex-direction: column;
+      gap: 2px;
+      padding-left: 16px;
+      border-left: 3px solid var(--border-primary);
+    }
+  }
+
+  &_topics {
+    .hint_trigger {
+      display: flex;
+      gap: 12px;
+      align-items: center;
+      padding: 10px 10px;
+      border-top: 1px solid var(--border-primary);
+      cursor: pointer;
+    }
+
+    .hint_content {
+      margin-left: 38px;
+    }
+  }
+
+  &_hints {
+    .hint_trigger {
+      display: flex;
+      gap: 12px;
+      align-items: center;
+      padding: 10px 10px;
+      border-top: 1px solid var(--border-primary);
+      cursor: pointer;
+    }
+
+    .hint_content {
+      margin-left: 38px;
+    }
+  }
+
+  &_footer {
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    align-self: self-end;
+
+    .like_dislike_container {
+      display: flex;
+      gap: 8px;
+    }
+
+    .stat_button {
+      display: flex;
+      gap: 8px;
+      align-items: center;
+      padding: 4px 6px;
+      border-radius: 4px;
+      background-color: var(--background-secondary);
+      color: var(--text-primary);
+      font-size: 14px;
+      font-weight: 500;
+      cursor: pointer;
+
+      &:hover {
+        background-color: var(--background-tertiary);
+      }
+    }
+  }
+}
