+++ src/routes/problem/components/details_pane/components/description/index.ts
@@ -1,0 +1,2 @@
+import Description from "./description";
+export default Description;
