+++ src/routes/problem/components/details_pane/components/editorial/editorial.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+type Props = {
+  data: any;
+  loading: boolean;
+};
+
+const Editorial = (props: Props) => {
+  return <div>Editorial</div>;
+};
+
+export default Editorial;
