+++ src/routes/problem/components/details_pane/components/editorial/index.ts
@@ -1,0 +1,2 @@
+import Editorial from "./editorial";
+export default Editorial;
