+++ src/routes/problem/components/details_pane/components/solutions/index.ts
@@ -1,0 +1,2 @@
+import Solutions from "./solutions";
+export default Solutions;
