+++ src/routes/problem/components/details_pane/components/solutions/solutions.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+type Props = {
+  data: any;
+  loading: boolean;
+};
+
+const Solutions = (props: Props) => {
+  return <div>Solutions</div>;
+};
+
+export default Solutions;
