+++ src/routes/problem/components/details_pane/details_pane.scss
@@ -2,171 +2,4 @@
   height: 100%;
   display: flex;
   flex-direction: column;
-  overflow: hidden;
-
-  &_content {
-    flex: 1;
-    overflow-y: auto;
-    padding: 1rem;
-
-    .problem_title {
-      font-size: 24px;
-      font-weight: 700;
-      margin-bottom: 8px;
-    }
-  }
-
-  &_metadata {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-    margin-bottom: 24px;
-  }
-
-  &_description {
-    margin-bottom: 48px;
-
-    h1,
-    h2,
-    h3,
-    h4 {
-      font-weight: 600;
-      margin: 18px 0 12px;
-    }
-
-    p {
-      margin-bottom: 12px;
-    }
-  }
-
-  &_examples {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin-bottom: 48px;
-
-    .example_heading {
-      font-weight: 700;
-      margin-bottom: 6px;
-    }
-
-    .example_fields_container {
-      display: flex;
-      flex-direction: column;
-      gap: 2px;
-      padding-left: 16px;
-      border-left: 3px solid var(--border-primary);
-    }
-  }
-
-  &_topics {
-    .hint_trigger {
-      display: flex;
-      gap: 12px;
-      align-items: center;
-      padding: 10px 10px;
-      border-top: 1px solid var(--border-primary);
-      cursor: pointer;
-    }
-
-    .hint_content {
-      margin-left: 38px;
-    }
-  }
-
-  &_hints {
-    .hint_trigger {
-      display: flex;
-      gap: 12px;
-      align-items: center;
-      padding: 10px 10px;
-      border-top: 1px solid var(--border-primary);
-      cursor: pointer;
-    }
-
-    .hint_content {
-      margin-left: 38px;
-    }
-  }
-
-  &_stats {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-    font-size: 0.875rem;
-    color: var(--text-secondary);
-
-    .stat-item {
-      display: flex;
-      align-items: center;
-      gap: 0.25rem;
-    }
-
-    .separator {
-      width: 1px;
-      height: 1rem;
-      background-color: #374151;
-    }
-
-    &--floating {
-      position: fixed;
-      bottom: 2rem;
-      left: 2rem;
-      display: flex;
-      gap: 1px;
-      padding: 2px;
-      border-radius: 8px;
-      background-color: var(--background-secondary);
-      box-shadow: 0 4px 12px -1px var(--shadow-color);
-      z-index: 10;
-
-      .stat-button {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        padding: 0.5rem 1rem;
-        background-color: var(--background-primary);
-        border: none;
-        color: var(--text-primary);
-        cursor: pointer;
-        transition: all 0.2s ease;
-
-        &:first-child {
-          border-top-left-radius: 6px;
-          border-bottom-left-radius: 6px;
-        }
-
-        &:last-child {
-          border-top-right-radius: 6px;
-          border-bottom-right-radius: 6px;
-        }
-
-        &:hover {
-          background-color: var(--background-hover);
-        }
-
-        &:active {
-          transform: scale(0.95);
-        }
-
-        .icon {
-          color: var(--text-secondary);
-        }
-
-        &--liked .icon {
-          color: #3b82f6;
-        }
-
-        &--disliked .icon {
-          color: #ef4444;
-        }
-
-        .count {
-          font-weight: 500;
-          font-size: 0.875rem;
-        }
-      }
-    }
-  }
 }
